add: member schema, page schema, modified: errors , routing and fixed overall errors

+++ server/controllers/postController.js
@@ -5,72 +5,82 @@ class PostController {
     this.postDataHandler = postDataHandler;
   }
 
-  getAllByType({ type }) {
-    var result, data;
+  async getPostsByType(type) {
+    let result, data;
     try {
-      data = this.postDataHandler.getPosts({ type: type });
-      result = {
-        type: type,
-        data: data,
-      }
+      //get data
+      data = await this.postDataHandler.getPostsByType(type);
+      //console.log(data[0], type)
+      result = { type, data };
     } catch (err) {
       throw new HttpExceptionHandler(400, err);
     }
     return result;
   }
 
-  getPost({ postId }) {
-    var result, data;
+  async getPost({ postId }) {
+    let result, data;
     try {
-      data = this.postDataHandler.getPost(postId);
-      result = {
-        type: data.type,
-        data: data,
-      }
+      //get data
+      data = await this.postDataHandler.getPostById(postId);
+      result = { type: data.type, data }
     } catch (err) {
       throw new HttpExceptionHandler(400, err);
     }
     return result;
   }
 
-  createPost({ post }) {
-    var result, data;
+  async createPost(user, postItem) {
+    let result, postId, data;
     try {
-      data = this.postDataHandler.createPost(post);
-      result= {
-        type: data.type,
-        data: data
-      }
+      //add author info from user info
+      postItem.authorId = user.id,
+      postItem.authorName = user.name
+      
+      //craete new post
+      postId = await this.postDataHandler.createPost(postItem);
+      data = await this.postDataHandler.getPostById(postId);
+      result = { type: data.type, data };
+      console.log(data)
     } catch (err) {
       throw new HttpExceptionHandler(400, err);
     }
     return result;
   }
 
-  updatePost(userId, { postId, updates }) {
-    var result, data;
+  async updatePost(user, { postId, updates }) {
+    let result, data;
     try {
-      this.postDataHandler.updatePost(userId, postId, updates);
-      data = this.postDataHandler.getPost(postId);
-      result = {
-        type: data.type,
-        data: data
+      //check permission
+      let { userId, role } = user;
+      let { authorId, type } = await this.postDataHandler.getPostById(postId);
+      if (!role === 'admin' || !authorId || !authorId === userId) {
+        throw "Cannot update";
       }
+      //update post
+      await this.postDataHandler.updatePost(postId, updates);
+      data = await this.postDataHandler.getPost(postId);
+      result = { type, data };
     } catch (err) {
       throw new HttpExceptionHandler(400, err);
     }
     return result;
   }
 
-  deletePost(userId, { postId }) {
-    var result, data;
+  async deletePost(user, { postId }) {
+    let result, data, post;
     try {
-      data = this.postDataHandler.getPost(postId);
-      this.postDataHandler.deletePost(userId, postId);
-      result = {
-        type: data.type,
-        data: data
+      //check permission
+      post = await this.postDataHandler.getPostById(postId);
+      let { userId, role } = user;
+      let { authorId } = post;
+      if (!role === 'admin' || !authorId || !authorId === userId) {
+        throw "Cannot delete"
       }
+      //delete post and return origianl post
+      data = await this.postDataHandler.getPost(postId);
+      await this.postDataHandler.deletePost(userId, postId);
+      result = { type, data };
     } catch (err) {
       throw new HttpExceptionHandler(400, err);
     }
